app/requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
huggingface_hub
psycopg2-binary
sqlalchemy
